pyproject.toml
[tool.poetry]
name = "synda"
version = "0.1.0"
description = "A synthetic data generator pipeline"
authors = ["TimothePearce <timothe.pearce@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
keywords = ["synthetic data", "pipeline", "llm"]
packages = [{include = "synda"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
pydantic = "^2.10.5"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
litellm = "1.55.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
nebula = "synda.cli:main"
